add gtid function (#4)
* 1.  add gtid command & filter execed command function
2.  command change :set => gtid set
3.  gtid command  exec =>gtid  exec
4.  add ctrip mode (normal, gtid)
5.  gtid expire => gtid expireat, when save aof file
6.  when run gtid exec command and gtid is execed ，must clean multi queue
7.  propagate gtid expire when key expire
8.  add gtid.lwm command ,for example（gtid.lwm A 100)
9.  add gtid info save rdb and load rdb
10. gtid command support comment
11. add gtid.auto command
12. add command (ctrip.merge_start, ctrip.merge , ctrip.merge_end)

* remove gtid_in_exec & use server.in_exec

* remove gtid + setex code when aof write file

* 1. ctrip_gtid project support -m32
2. hotfix test

* remove ci test ubuntu-old(16)

+++ src/db.c
@@ -1463,8 +1463,12 @@ void deleteExpiredKeyAndPropagate(redisDb *db, robj *keyobj) {
  * will be consistent even if we allow write operations against expiring
  * keys. */
 void propagateExpire(redisDb *db, robj *key, int lazy) {
+    if(isGtidEnabled()) {
+        propagateGtidExpire(db, key, lazy);
+        return;
+    }
     robj *argv[2];
-
+    
     argv[0] = lazy ? shared.unlink : shared.del;
     argv[1] = key;
     incrRefCount(argv[0]);
